Add parallel implementation of Negamax.
The effectiveness of this can vary considerably, depending upon: the game
properties, such as the branching factor; the number of threads specified
for the pool; the number of hardware threads; and the kind of hardware.
For instance, on my desktop with an i7 2600K running Arch, the benchmarks
show a speedup of about 2.7x; while, on my server with 2x Xeon 2630 v3 and
running Arch under a SmartOS KVM, there's a 6x slowdown! Even bringing the
threads down from `num_cpus::get()` to just 1 results in a 1.3x slowdown.
In other words, this is by no means a "silver bullet."