Pass a Vec in the generate_moves API.
This simplifies dealing with moves on both sides of the API and
prevents out of bounds for games with a high branching factor.
Backwards incompatible, so bumping the version to 0.2.0.

The original reason for using a stack-allocated array of optional
moves was to prevent allocations in the inner loop. However, it
turns out to be easy to just cache all the Vecs and their backing
storage, so the perf actually improves (around 1.1x), as you don't
have to repeatedly set large chunks of the stack to the non-zero
Option<M>.